requirements.txt
streamlit==1.31.0
playwright==1.41.0
requests==2.31.0
beautifulsoup4==4.12.3
sqlalchemy==2.0.25
pandas==2.2.0
python-dateutil==2.8.2
plotly==5.18.0
lxml==5.1.0
